requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.0.10
django-cors-headers==3.10.1
django-environ==0.10.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.4
psycopg2==2.9.7
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3
whitenoise==6.0.0
